#1615 project reorganized to make the gwt module a usable artifact
* GWT plugin changed to mojo's GWT plugin to have a jar project type (prevoius `gwt-lib` type was not picked up by other plugins when building the final artifact),
* all super-sources moved to the GWT module,
* potential Guava conflicts resolved by merging Guava's Executors impl.
