--- javaslang-gwt/src/main/java/JavaslangExample.gwt.xml
@@ -1,4 +1,0 @@
-<module>
-    <inherits name='Javaslang'/>
-    <entry-point class="javaslang.JavaslangEntryPoint"/>
-</module>
