+++ javaslang-gwt/src/main/resources/super/java/lang/LinkageError.java
