+++ javaslang-gwt/src/main/resources/super/java/lang/ThreadDeath.java
