+++ javaslang-gwt/src/main/resources/super/java/lang/VirtualMachineError.java
