+++ javaslang-gwt/src/main/resources/super/java/util/concurrent/CancellationException.java
