+++ javaslang-gwt/src/main/resources/super/java/util/concurrent/ExecutorService.java
@@ -1,0 +1,14 @@
+/*     / \____  _    _  ____   ______  / \ ____  __    _______
+ *    /  /    \/ \  / \/    \ /  /\__\/  //    \/  \  //  /\__\   JΛVΛSLΛNG
+ *  _/  /  /\  \  \/  /  /\  \\__\\  \  //  /\  \ /\\/ \ /__\ \   Copyright 2014-2016 Javaslang, http://javaslang.io
+ * /___/\_/  \_/\____/\_/  \_/\__\/__/\__\_/  \_//  \__/\_____/   Licensed under the Apache License, Version 2.0
+ */
+package java.util.concurrent;
+
+/**
+ * GWT emulated version of the {@link ExecutorService} with a limited set of methods used by Javaslang.
+ */
+public interface ExecutorService extends Executor {
+
+    Future<?> submit(Runnable task);
+}
