+++ javaslang-gwt/src/main/resources/super/java/util/concurrent/Executors.java
@@ -1,0 +1,52 @@
+/*     / \____  _    _  ____   ______  / \ ____  __    _______
+ *    /  /    \/ \  / \/    \ /  /\__\/  //    \/  \  //  /\__\   JΛVΛSLΛNG
+ *  _/  /  /\  \  \/  /  /\  \\__\\  \  //  /\  \ /\\/ \ /__\ \   Copyright 2014-2016 Javaslang, http://javaslang.io
+ * /___/\_/  \_/\____/\_/  \_/\__\/__/\__\_/  \_//  \__/\_____/   Licensed under the Apache License, Version 2.0
+ */
+package java.util.concurrent;
+
+import javaslang.concurrent.CurrentThreadExecutorService;
+
+/**
+ * GWT emulated implementation of {@link Executors} merged with Guava implementation to address potential
+ * dependency conflicts.
+ */
+public class Executors {
+
+    public static <T> Callable<T> callable(Runnable task, T result) {
+        if (task == null) {
+            throw new NullPointerException();
+        }
+        return new RunnableAdapter<T>(task, result);
+    }
+
+    public static Callable<Object> callable(Runnable task) {
+        if (task == null) {
+            throw new NullPointerException();
+        }
+        return new RunnableAdapter<Object>(task, null);
+    }
+
+    public static ExecutorService newCachedThreadPool() {
+        return new CurrentThreadExecutorService();
+    }
+
+    static final class RunnableAdapter<T> implements Callable<T> {
+
+        final Runnable task;
+        final T result;
+
+        RunnableAdapter(Runnable task, T result) {
+            this.task = task;
+            this.result = result;
+        }
+
+        public T call() {
+            task.run();
+            return result;
+        }
+    }
+
+    private Executors() {
+    }
+}
