--- javaslang/src/main/java/Javaslang.gwt.xml
@@ -1,5 +1,0 @@
-<module>
-    <inherits name='com.google.gwt.user.User'/>
-    <super-source path="super" />
-    <source path="javaslang"/>
-</module>
