--- javaslang/src/main/resources/super/java/util/concurrent/atomic/AtomicInteger.java
@@ -1,325 +1,0 @@
-/*
- * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- */
-
-/*
- *
- *
- *
- *
- *
- * Written by Doug Lea with assistance from members of JCP JSR-166
- * Expert Group and released to the public domain, as explained at
- * http://creativecommons.org/publicdomain/zero/1.0/
- */
-
-package java.util.concurrent.atomic;
-import java.util.function.IntUnaryOperator;
-import java.util.function.IntBinaryOperator;
-
-/**
- * An {@code int} value that may be updated atomically.  See the
- * {@link java.util.concurrent.atomic} package specification for
- * description of the properties of atomic variables. An
- * {@code AtomicInteger} is used in applications such as atomically
- * incremented counters, and cannot be used as a replacement for an
- * {@link java.lang.Integer}. However, this class does extend
- * {@code Number} to allow uniform access by tools and utilities that
- * deal with numerically-based classes.
- *
- * @since 1.5
- * @author Doug Lea
- */
-public class AtomicInteger extends Number implements java.io.Serializable {
-    private static final long serialVersionUID = 6214790243416807050L;
-
-    private int value;
-
-    /**
-     * Creates a new AtomicInteger with the given initial value.
-     *
-     * @param initialValue the initial value
-     */
-    public AtomicInteger(int initialValue) {
-        value = initialValue;
-    }
-
-    /**
-     * Creates a new AtomicInteger with initial value {@code 0}.
-     */
-    public AtomicInteger() {
-    }
-
-    /**
-     * Gets the current value.
-     *
-     * @return the current value
-     */
-    public final int get() {
-        return value;
-    }
-
-    /**
-     * Sets to the given value.
-     *
-     * @param newValue the new value
-     */
-    public final void set(int newValue) {
-        value = newValue;
-    }
-
-    /**
-     * Eventually sets to the given value.
-     *
-     * @param newValue the new value
-     * @since 1.6
-     */
-    public final void lazySet(int newValue) {
-        set(newValue);
-    }
-
-    /**
-     * Atomically sets to the given value and returns the old value.
-     *
-     * @param newValue the new value
-     * @return the previous value
-     */
-    public final int getAndSet(int newValue) {
-        int old = value;
-        set(newValue);
-        return old;
-    }
-
-    /**
-     * Atomically sets the value to the given updated value
-     * if the current value {@code ==} the expected value.
-     *
-     * @param expect the expected value
-     * @param update the new value
-     * @return {@code true} if successful. False return indicates that
-     * the actual value was not equal to the expected value.
-     */
-    public final boolean compareAndSet(int expect, int update) {
-        if (expect == value) {
-            value = update;
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * Atomically sets the value to the given updated value
-     * if the current value {@code ==} the expected value.
-     *
-     * <p><a href="package-summary.html#weakCompareAndSet">May fail
-     * spuriously and does not provide ordering guarantees</a>, so is
-     * only rarely an appropriate alternative to {@code compareAndSet}.
-     *
-     * @param expect the expected value
-     * @param update the new value
-     * @return {@code true} if successful
-     */
-    public final boolean weakCompareAndSet(int expect, int update) {
-        return compareAndSet(expect, update);
-    }
-
-    /**
-     * Atomically increments by one the current value.
-     *
-     * @return the previous value
-     */
-    public final int getAndIncrement() {
-        int old = value;
-        value++;
-        return old;
-    }
-
-    /**
-     * Atomically decrements by one the current value.
-     *
-     * @return the previous value
-     */
-    public final int getAndDecrement() {
-        int old = value;
-        value--;
-        return old;
-    }
-
-    /**
-     * Atomically adds the given value to the current value.
-     *
-     * @param delta the value to add
-     * @return the previous value
-     */
-    public final int getAndAdd(int delta) {
-        int old = value;
-        value += delta;
-        return old;
-    }
-
-    /**
-     * Atomically increments by one the current value.
-     *
-     * @return the updated value
-     */
-    public final int incrementAndGet() {
-        value++;
-        return value;
-    }
-
-    /**
-     * Atomically decrements by one the current value.
-     *
-     * @return the updated value
-     */
-    public final int decrementAndGet() {
-        value--;
-        return value;
-    }
-
-    /**
-     * Atomically adds the given value to the current value.
-     *
-     * @param delta the value to add
-     * @return the updated value
-     */
-    public final int addAndGet(int delta) {
-        value += delta;
-        return value;
-    }
-
-    /**
-     * Atomically updates the current value with the results of
-     * applying the given function, returning the previous value. The
-     * function should be side-effect-free, since it may be re-applied
-     * when attempted updates fail due to contention among threads.
-     *
-     * @param updateFunction a side-effect-free function
-     * @return the previous value
-     * @since 1.8
-     */
-    public final int getAndUpdate(IntUnaryOperator updateFunction) {
-        int old = value;
-        value = updateFunction.applyAsInt(value);
-        return old;
-    }
-
-    /**
-     * Atomically updates the current value with the results of
-     * applying the given function, returning the updated value. The
-     * function should be side-effect-free, since it may be re-applied
-     * when attempted updates fail due to contention among threads.
-     *
-     * @param updateFunction a side-effect-free function
-     * @return the updated value
-     * @since 1.8
-     */
-    public final int updateAndGet(IntUnaryOperator updateFunction) {
-        value = updateFunction.applyAsInt(value);
-        return value;
-    }
-
-    /**
-     * Atomically updates the current value with the results of
-     * applying the given function to the current and given values,
-     * returning the previous value. The function should be
-     * side-effect-free, since it may be re-applied when attempted
-     * updates fail due to contention among threads.  The function
-     * is applied with the current value as its first argument,
-     * and the given update as the second argument.
-     *
-     * @param x the update value
-     * @param accumulatorFunction a side-effect-free function of two arguments
-     * @return the previous value
-     * @since 1.8
-     */
-    public final int getAndAccumulate(int x,
-            IntBinaryOperator accumulatorFunction) {
-        int old = value;
-        value = accumulatorFunction.applyAsInt(value, x);
-        return old;
-    }
-
-    /**
-     * Atomically updates the current value with the results of
-     * applying the given function to the current and given values,
-     * returning the updated value. The function should be
-     * side-effect-free, since it may be re-applied when attempted
-     * updates fail due to contention among threads.  The function
-     * is applied with the current value as its first argument,
-     * and the given update as the second argument.
-     *
-     * @param x the update value
-     * @param accumulatorFunction a side-effect-free function of two arguments
-     * @return the updated value
-     * @since 1.8
-     */
-    public final int accumulateAndGet(int x,
-            IntBinaryOperator accumulatorFunction) {
-        value = accumulatorFunction.applyAsInt(value, x);
-        return value;
-    }
-
-    /**
-     * Returns the String representation of the current value.
-     * @return the String representation of the current value
-     */
-    public String toString() {
-        return Integer.toString(get());
-    }
-
-    /**
-     * Returns the value of this {@code AtomicInteger} as an {@code int}.
-     */
-    public int intValue() {
-        return get();
-    }
-
-    /**
-     * Returns the value of this {@code AtomicInteger} as a {@code long}
-     * after a widening primitive conversion.
-     * @jls 5.1.2 Widening Primitive Conversions
-     */
-    public long longValue() {
-        return (long)get();
-    }
-
-    /**
-     * Returns the value of this {@code AtomicInteger} as a {@code float}
-     * after a widening primitive conversion.
-     * @jls 5.1.2 Widening Primitive Conversions
-     */
-    public float floatValue() {
-        return (float)get();
-    }
-
-    /**
-     * Returns the value of this {@code AtomicInteger} as a {@code double}
-     * after a widening primitive conversion.
-     * @jls 5.1.2 Widening Primitive Conversions
-     */
-    public double doubleValue() {
-        return (double)get();
-    }
-
-}
