+++ pom.xml
@@ -114,16 +114,6 @@ Note: The maven build currently needs to be started from the javaslang root dir
                 <artifactId>jmh-generator-annprocess</artifactId>
                 <version>${jmh.version}</version>
             </dependency>
-            <dependency>
-                <groupId>com.google.gwt</groupId>
-                <artifactId>gwt-user</artifactId>
-                <version>${gwt.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.google.gwt</groupId>
-                <artifactId>gwt-dev</artifactId>
-                <version>${gwt.version}</version>
-            </dependency>
         </dependencies>
     </dependencyManagement>
     <build>
@@ -297,11 +287,6 @@ Note: The maven build currently needs to be started from the javaslang root dir
                             </goals>
                         </execution>
                     </executions>
-                </plugin>
-                <plugin>
-                    <groupId>net.ltgt.gwt.maven</groupId>
-                    <artifactId>gwt-maven-plugin</artifactId>
-                    <version>${maven.gwt.plugin}</version>
                 </plugin>
             </plugins>
         </pluginManagement>
